Deploys to testpypi when builds are not tagged
*  Updates .travis.yml with the configuration data for testpypi
*  Adds the script, `travis_set_build.py`, which will update the
   __version__ string with the Travis build number

The idea is to use the `skip` param of the new script in conjunction
with the $TRAVIS_TAG variable to control whether to update the
__version__ string with the Travis build number. If the build is
not a tag, it should be deployed to testpypi, with a build version.
If the build is a tag, it should be deployed to production pypi and
the version should be updated outside of Travis.